requirements.txt
anyio==3.6.2
asgiref==3.6.0
click==8.1.3
fastapi==0.89.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
peewee==3.15.4
pydantic==1.10.4
Pygments==2.13.0
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
sqlite-web==0.4.1
sqlparse==0.4.3
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
zipp==3.11.0
sqlalchemy==2.0.0
psycopg2==2.9.5
passlib==1.7.4
bcrypt==4.0.1
ecdsa==0.18.0
pyasn1==0.4.8
python-jose==3.3.0
rsa==4.9
six==1.16.0

python-multipart==0.0.5
boto3==1.26.69
twilio==8.2.1